#23: add select currency drawer test screen
